Clamp illegal negative array sizes to 0.
In debug mode, negative sizes are caught by a debug_assert.
But in release mode they result in integer overflow and confusing
bounds-check panic messages. This happens when code simply declares
a dummy array argument, even if it's never accessed.

To avoid that confusion, without the cost of release-mode asserts here,
clamp to 0. (It will still fail a bounds-check if it's accessed.)